#73 FEAT - 배포 실패 시, 이전 배포 상태 유지

+++ scripts/deploy.sh
@@ -26,16 +26,26 @@ if [ -z "$EXIST_BLUE" ]; then
   # 30초 동안 대기
   sleep 30
 
-  # /home/ubuntu/deploy.log: 로그 파일에 "green 중단 시작"이라는 내용을 추가
-  echo "green 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
+  # blue가 문제 없이 배포 되었는지 현재 실행여부를 확인한다
+  BLUE_HEALTH=$(sudo docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml ps | grep Up)
 
-  # docker-compose.green.yml 파일을 사용하여 spring-green 프로젝트의 컨테이너를 중지
-  sudo docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml down
+  # blue가 현재 실행중이지 않다면 -> 런타임 에러 또는 다른 이유로 배포가 되지 못한 상태
+  if [ -z "$BLUE_HEALTH" ]; then
+    continue
+  # blue가 현재 실행되고 있는 경우에만 green을 종료
+  else
 
-   # 사용하지 않는 이미지 삭제
-  sudo docker image prune -af
+    # /home/ubuntu/deploy.log: 로그 파일에 "green 중단 시작"이라는 내용을 추가
+    echo "green 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
 
-  echo "green 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
+    # docker-compose.green.yml 파일을 사용하여 spring-green 프로젝트의 컨테이너를 중지
+    sudo docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml down
+
+    # 사용하지 않는 이미지 삭제
+    sudo docker image prune -af
+
+    echo "green 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
+  fi
 
 # blue가 실행중이면 green up
 else
@@ -44,14 +54,21 @@ else
 
   sleep 30
 
-  echo "blue 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
-  sudo docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml down
-  sudo docker image prune -af
+  GREEN_HEALTH=$(sudo docker-compose -p ${DOCKER_APP_NAME}-green -f docker-compose.green.yml ps | grep Up)
 
-  echo "blue 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
+  if [ -z "$GREEN_HEALTH" ]; then
+    continue
+  else
 
+      # /home/ubuntu/deploy.log: 로그 파일에 "blue 중단 시작"이라는 내용을 추가
+      echo "blue 중단 시작 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
+
+      # docker-compose.blue.yml 파일을 사용하여 spring-green 프로젝트의 컨테이너를 중지
+      sudo docker-compose -p ${DOCKER_APP_NAME}-blue -f docker-compose.blue.yml down
+
+      # 사용하지 않는 이미지 삭제
+      sudo docker image prune -af
+
+      echo "blue 중단 완료 : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
+  fi
 fi
-  echo "배포 종료  : $(date +%Y)-$(date +%m)-$(date +%d) $(date +%H):$(date +%M):$(date +%S)" >> /home/ubuntu/deploy.log
-
-  echo "===================== 배포 완료 =====================" >> /home/ubuntu/deploy.log
-  echo >> /home/ubuntu/deploy.log
